+++ src/js/00-core.js
@@ -3,7 +3,7 @@
 
   var UI = window.UI || {};
 
-  UI.version = "1.9.0";
+  UI.version = "1.10.0";
   UI._initializers = UI._initializers || [];
 
   UI.q = function (selector, root) {